pc5/requirements.txt
psycopg2-binary==2.9.5
redis==4.5.4
pika==1.3.1
requests==2.28.2
aiohttp==3.8.4
python-dotenv==1.0.0
prometheus-client==0.16.0
fastapi==0.68.1
uvicorn==0.15.0
python-json-logger==2.0.7
backoff==2.2.1
beautifulsoup4==4.9.3
lxml==4.9.0